+++ Documents/Documents/NSString+WordCount.h
@@ -15,7 +15,7 @@ NS_ASSUME_NONNULL_BEGIN
 
 @interface NSString (WordCount)
 
-- (int)checkWordCount;
+- (NSUInteger)checkWordCount;
 
 @end
